Rename the App component to Root
